+++ docs/release_notes.rst
@@ -9,7 +9,7 @@ Release Notes
 * Added ``equals()`` and ``strictEquals()`` and ``sameValue()`` to ``IV8Value``
 * Added ``getIdentityHash()`` to ``IV8ValueReference``
 * Added ``isDead()``, ``isInUse()``, ``callAsConstructor()`` and ``terminateExecution()`` to ``V8Runtime``
-* Added V8 typed array
+* Added V8 typed array and data view
 * Added ``IJavetEngineGuard``
 
 0.7.1